Split dep tests into separate file.
